requirements_freeze.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioshutil==1.5
aiosignal==1.4.0
airportsdata==20250811
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anthropic==0.66.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.3
backoff==2.2.1
beautifulsoup4==4.13.5
black==25.1.0
blobfile==3.0.0
boto3==1.40.25
botocore==1.40.25
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
compressed-tensors==0.11.0
contourpy==1.3.3
cryptography==45.0.7
cuda-bindings==13.0.1
cuda-pathfinder==1.2.1
cuda-python==13.0.1
cycler==0.12.1
Cython==3.1.3
dataclasses-json==0.6.7
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
doclayout_yolo==0.0.4
effdet==0.4.1
einops==0.8.1
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.1
fast-langdetect==0.2.5
fastapi==0.116.1
fasttext-predict==0.9.2.4
fastuuid==0.12.0
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
fire==0.7.1
flashinfer-python==0.2.9rc2
flatbuffers==25.2.10
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
google-genai==1.33.0
googleapis-common-protos==1.70.0
gradio==5.44.1
gradio_client==1.12.1
gradio_pdf==0.0.22
groovy==0.1.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.9
hf_transfer==0.1.9
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
interegular==0.3.3
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
json_repair==0.50.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langdetect==1.0.9
lark==1.2.2
lazy_loader==0.4
litellm==1.76.3
llguidance==0.7.30
loguru==0.7.3
lxml==6.0.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mineru==2.2.1
ml_dtypes==0.5.3
modelscope==1.29.2
mpmath==1.3.0
msgspec==0.19.0
msoffcrypto-tool==5.4.2
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.14.1
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1
openai==1.106.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.3
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
paddlepaddle-gpu==2.6.2
pandas==2.3.2
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250506
pdftext==0.6.3
pexpect==4.9.0
pi_heif==1.1.0
pikepdf==9.10.2
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-zerox==0.0.7
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pynvml==13.0.1
pypandoc==1.15
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
RapidFuzz==3.14.0
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
robust-downloader==0.0.2
rpds-py==0.27.1
rsa==4.9.1
ruff==0.12.12
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
setproctitle==1.3.7
setuptools==80.9.0
sgl-kernel==0.2.8
sglang==0.4.10.post2
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.47.3
stringzilla==3.12.6
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
thop==0.1.1.post2209072238
tifffile==2025.8.28
tiktoken==0.11.0
timm==1.0.16
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torch_memory_saver==0.0.8
torchao==0.9.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
triton==3.4.0
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.195
ultralytics-thop==2.0.17
unstructured==0.18.14
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
unstructured_paddleocr==2.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
weaviate-client==4.16.9
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
xgrammar==0.1.22
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
